feat(dioxus-translate-macro): add tracing dependencies and move tests to separate file
